+++ changelog.md
@@ -7,6 +7,7 @@
 - Added "Optimize Linear Cast" option, mitigating lag caused by having lots of very fast moving projectiles on the screen
 - Added "Unpause / Repause Fix"
 - Further improved "Optimize Bullet Trails" option to save even more RAM
+- Added "Optimize Linear Cast Pool" to fix a memory leak in physics calculations
 
 ## 1.3.3 (2025-06-01)